requirements.txt
boto3==1.5.8
docopt==0.6.2
python-dateutil==2.6.1
pytz==2017.3
